Add /one-time-payment POST endpoint
Currently, we are in need of a way for API customers to send a payment. This change adds POST /one-time-payment which takes a JSON object and returns JSON with the new balance and the due date for the next payment (currently 15 days in the future). Now that we have the base laid for this feature, we can think about adding logic for our Match program and making it so the next due date is never on a weekend (because what a terrible weekend).